Rename classes and package
