+++ src/main/java/org/complate/spring/boot/sample/Application.java
@@ -1,13 +1,12 @@
-package com.github.complate.springdemo;
+package org.complate.spring.boot.sample;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class SpringdemoApplication {
+public class Application {
 
     public static void main(String[] args) {
-        SpringApplication.run(SpringdemoApplication.class, args);
+        SpringApplication.run(Application.class, args);
     }
-
 }
