+++ src/main/java/org/complate/spring/boot/sample/ComplateConfiguration.java
@@ -1,11 +1,12 @@
-package com.github.complate.springdemo;
+package org.complate.spring.boot.sample;
 
 import com.github.complate.ComplateViewResolver;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-public class SpringdemoConfiguration {
+public class ComplateConfiguration {
+
     @Bean
     public ComplateViewResolver complateViewResolver() {
         return new ComplateViewResolver();
